+++ R/lsa.R
@@ -1,12 +1,13 @@
 #' Latent Semantic Analysis model
 #' 
 #' Train a Latent Semantic Analysis model (Deerwester et al., 1990) on a [quanteda::tokens] object.
-#' @param x a [quanteda::tokens] object.
+#' @param x a [quanteda::tokens] or [quanteda::tokens_xptr] object.
 #' @param dim the size of the word vectors.
 #' @param min_count the minimum frequency of the words. Words less frequent than 
 #'    this in `x` are removed before training.
 #' @param engine select the engine perform SVD to generate word vectors.
 #' @param weight weighting scheme passed to [quanteda::dfm_weight()]. 
+#' @param tolower if `TRUE` lower-case all the tokens before fitting the model.
 #' @param verbose if `TRUE`, print the progress of training.
 #' @param ... additional arguments.
 #' @returns Returns a textmodel_wordvector object with the following elements:
@@ -15,6 +16,7 @@
 #'   \item{frequency}{the frequency of words in `x`.}
 #'   \item{engine}{the SVD engine used.}
 #'   \item{weight}{weighting scheme.}
+#'   \item{min_count}{the value of min_count.}
 #'   \item{concatenator}{the concatenator in `x`.}
 #'   \item{call}{the command used to execute the function.}
 #'   \item{version}{the version of the wordvector package.}
@@ -45,7 +47,7 @@
 #' }
 textmodel_lsa <- function(x, dim = 50, min_count = 5L, 
                           engine = c("RSpectra", "irlba", "rsvd"), 
-                          weight = "count", verbose = FALSE, ...) {
+                          weight = "count", tolower = TRUE, verbose = FALSE, ...) {
     UseMethod("textmodel_lsa")   
 }
 
@@ -54,9 +56,9 @@ textmodel_lsa <- function(x, dim = 50, min_count = 5L,
 #' @method textmodel_lsa tokens
 textmodel_lsa.tokens <- function(x, dim = 50L, min_count = 5L, 
                                  engine = c("RSpectra", "irlba", "rsvd"), 
-                                 weight = "count", verbose = FALSE, ...) {
+                                 weight = "count", tolower = TRUE, verbose = FALSE, ...) {
     
-    result <- textmodel_lsa(dfm(x, remove_padding = TRUE), 
+    result <- textmodel_lsa(dfm(x, remove_padding = TRUE, tolower = tolower), 
                             dim = dim, min_count = min_count, engine = engine, weight = weight,
                             verbose = verbose, ...)
     result$call = try(match.call(sys.function(-1), call = sys.call(-1)), silent = TRUE)
@@ -68,7 +70,7 @@ textmodel_lsa.tokens <- function(x, dim = 50L, min_count = 5L,
 #' @method textmodel_lsa dfm
 textmodel_lsa.dfm <- function(x, dim = 50L, min_count = 5L, 
                                  engine = c("RSpectra", "irlba", "rsvd"), 
-                                 weight = "count", verbose = FALSE, ...) {
+                                 weight = "count", tolower = TRUE, verbose = FALSE, ...) {
     
     engine <- match.arg(engine)
     dim <- check_integer(dim, min = 2)
@@ -90,10 +92,10 @@ textmodel_lsa.dfm <- function(x, dim = 50L, min_count = 5L,
     result <- list(
         values = wov,
         dim = dim,
-        min_count = min_count,
         frequency = featfreq(x),
         engine = engine,
         weight = weight,
+        min_count = min_count,
         concatenator = meta(x, field = "concatenator", type = "object"),
         call = try(match.call(sys.function(-1), call = sys.call(-1)), silent = TRUE),
         version = utils::packageVersion("wordvector")