+++ R/word2vec.R
@@ -1,7 +1,7 @@
 #' Word2vec model
 #' 
 #' Train a Word2vec model (Mikolov et al., 2023) in different architectures on a [quanteda::tokens] object.
-#' @param x a [quanteda::tokens] object.
+#' @param x a [quanteda::tokens] or [quanteda::tokens_xptr] object.
 #' @param dim the size of the word vectors.
 #' @param type the architecture of the model; either "cbow" (continuous back of words) or "skip-gram".
 #' @param min_count the minimum frequency of the words. Words less frequent than 
@@ -16,6 +16,7 @@
 #' @param sample the rate of sampling of words based on their frequency. Sampling is 
 #'   disabled when `sample = 1.0`
 #' @param normalize if `TRUE`, normalize the vectors in `values` and `weights`.
+#' @param tolower lower-case all the tokens before fitting the model.
 #' @param verbose if `TRUE`, print the progress of training.
 #' @param ... additional arguments.
 #' @returns Returns a textmodel_wordvector object with the following elements:
@@ -29,6 +30,7 @@
 #'   \item{alpha}{the initial learning rate.}
 #'   \item{use_ns}{the use of negative sampling.}
 #'   \item{ns_size}{the size of negative samples.}
+#'   \item{min_count}{the value of min_count.}
 #'   \item{concatenator}{the concatenator in `x`.}
 #'   \item{call}{the command used to execute the function.}
 #'   \item{version}{the version of the wordvector package.}
@@ -65,7 +67,7 @@
 textmodel_word2vec <- function(x, dim = 50, type = c("cbow", "skip-gram"), 
                                min_count = 5L, window = ifelse(type == "cbow", 5L, 10L), 
                                iter = 10L, alpha = 0.05, use_ns = TRUE, ns_size = 5L, 
-                               sample = 0.001, normalize = TRUE,
+                               sample = 0.001, normalize = TRUE, tolower = TRUE,
                                verbose = FALSE, ...) {
     UseMethod("textmodel_word2vec")
 }
@@ -77,7 +79,7 @@ textmodel_word2vec <- function(x, dim = 50, type = c("cbow", "skip-gram"),
 textmodel_word2vec.tokens <- function(x, dim = 50L, type = c("cbow", "skip-gram"), 
                                       min_count = 5L, window = ifelse(type == "cbow", 5L, 10L), 
                                       iter = 10L, alpha = 0.05, use_ns = TRUE, ns_size = 5L, 
-                                      sample = 0.001, normalize = TRUE,
+                                      sample = 0.001, normalize = TRUE, tolower = TRUE,
                                       verbose = FALSE, ..., old = FALSE) {
     
     type <- match.arg(type)
@@ -90,23 +92,26 @@ textmodel_word2vec.tokens <- function(x, dim = 50L, type = c("cbow", "skip-gram"
     alpha <- check_double(alpha, min = 0)
     sample <- check_double(sample, min = 0)
     normalize <- check_logical(normalize)
+    tolower <- check_logical(tolower)
     verbose <- check_logical(verbose)
 
     type <- match(type, c("cbow", "skip-gram"))
     if (old)
         type <- type * 10
     
-    # NOTE: use tokens_xptr?
+    x <- as.tokens_xptr(x)
+    if (tolower)
+        x <- tokens_tolower(x)
     x <- tokens_trim(x, min_termfreq = min_count, termfreq_type = "count")
-    result <- cpp_w2v(as.tokens(x), attr(x, "types"), 
-                      minWordFreq = min_count,
-                      size = dim, window = window,
+    
+    result <- cpp_w2v(x, size = dim, window = window,
                       sample = sample, withHS = !use_ns, negative = ns_size, 
                       threads = get_threads(), iterations = iter,
                       alpha = alpha, type = type, normalize = normalize, verbose = verbose)
     if (!is.null(result$message))
         stop("Failed to train word2vec (", result$message, ")")
-
+    
+    result$min_count <- min_count
     result$concatenator <- meta(x, field = "concatenator", type = "object")
     result$call <- try(match.call(sys.function(-1), call = sys.call(-1)), silent = TRUE)
     result$version <- utils::packageVersion("wordvector")